requirements.txt
pandas==2.2.3
openpyxl==3.1.5
matplotlib==3.10.1

--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.5.0

pyyaml==6.0.2
einops==0.8.1
torcheval==0.0.7
seaborn==0.13.2